# packages/shared/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: workspace-chat API
  version: 0.0.0
  description: >-
    エラーの応答の本体はすべて ErrorResponse である。仕様に無いパス（前置き /api の外を含む）は
    404（not_found）を返し、パスごとには載せない。
servers:
  # 全ルートの前置きは /api（#77 の決定。apps/api/src/app-setup.ts の createApp が付ける）。
  - url: /api
# 既定ですべての操作にアクセストークンを求める。認証を要さない操作だけが `security: []` で外す
# （api 側は auth/access-token.guard.ts が同じ既定を持つ。一致は auth/access-token.test.ts が仕様を読んで確かめる）。
security:
  - bearerAuth: []
paths:
...
  /workspaces/{id}/channels/{channelId}/attachments/uploads:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/ChannelId'
    post:
      operationId: createAttachmentUpload
      summary: 添付ファイルのアップロード用の署名付き URL の発行（F-27・F-28）
      description: >-
        チャンネルの添付を上げる署名付き URL を発行する（機能一覧 11.1）。参加者だけが発行できる。コードは2段階: 所属していなければ種別によらず 404。
        所属していて参加していなければ、パブリックは 403 not_a_channel_member・プライベートは 404（オーナーでも参加していなければ同じ）。
        URL の宛先は隔離用のキー quarantine/workspace/{id}/channel/{channelId}/{uploadId}/{保存名} であり、キーはすべてサーバーが組み立てる
        （保存名はファイル名の英数字・.・_・- 以外と先頭の . を _ に置き換え、キー全体が 1,024 バイトに収まるよう切り詰めたもの）。
        有効期限は 5 分で、Content-Type と If-None-Match: * を署名に含む。
        申告した Content-Type が許可リスト（画像 jpeg / png / gif / webp、動画 mp4 / webm、文書 pdf / txt / csv / md / docx / xlsx / pptx、圧縮 zip）に無ければ
        422 unsupported_file_type、種別の上限（画像 10 MB・動画 100 MB・文書と圧縮 25 MB）を超えれば 422 file_too_large。
        申告は信用せず、確定で中身から検証する。利用者単位で 10 分に 30 回まで
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequest'
      responses:
        '201':
          description: 発行したアップロード
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadTicket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: パブリックチャンネルに参加していない（not_a_channel_member）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: 受け付けない形式（unsupported_file_type）か、上限を超える大きさ（file_too_large）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workspaces/{id}/channels/{channelId}/attachments/uploads/{uploadId}/complete:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/ChannelId'
      - $ref: '#/components/parameters/UploadId'
    post:
      operationId: completeAttachmentUpload
      summary: 添付ファイルのアップロードの確定（F-27・F-28）
      description: >-
        本人に払い出され、このワークスペースとチャンネルで発行した識別子だけを確定できる（そうでなければ 404。識別子の存在を認めず、確定の権利も使わない）。
        キーとやり直す参加者判定の対象チャンネルは、発行のときにサーバーが保持した値から組み立てる（機能一覧 11.1）。
        確定の時点で参加者判定をやり直し、落ちたら発行と同じ2段階のコード（404 / 403 not_a_channel_member）を返して隔離用のキーを削除する。
        通れば中身から形式を検証し、検証した版に固定して配信用のキー workspace/{id}/channel/{channelId}/{uploadId}/{保存名} へコピーし、隔離用のキーを削除する
        （保存名の拡張子は検証した形式のものに付け替え、配信の Content-Type と Content-Disposition は検証した形式から決める）。
        検証に通らなければ配信用のキーへ移さず隔離用のキーを削除し、422 を返す——許可リストに無い・中身が形式に合わない（unsupported_file_type）、
        種別の上限を超える（file_too_large）、本体が PUT されていない（upload_not_received）。
        確定は識別子ごとに1回だけ行い、2回目以降はコピーも削除もやり直さず、1回目と同じ結果を返す（参加者判定に落ちたときも同じ。やり直すには発行からやり直す）。
        確定の途中にもう1つ確定を求めると 409 upload_in_progress。確定した添付は、投稿（postMessage）の attachmentIds で投稿に付ける。
        利用者単位で 10 分に 30 回まで
      responses:
        '200':
          description: 確定した添付
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: パブリックチャンネルに参加していない（not_a_channel_member）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '409':
          description: 同じ識別子の確定が進行中（upload_in_progress）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 検証に通らなかった（unsupported_file_type・file_too_large・upload_not_received）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      description: ログイン（/auth/login）とリフレッシュ（/auth/refresh）が返すアクセストークン（JWT・HS256）
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      description: チャンネルの id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
...
    UploadId:
      name: uploadId
      in: path
      required: true
      description: アップロードの識別子（発行の応答の uploadId）。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
...
    WorkspaceId:
      name: id
      in: path
      required: true
      description: ワークスペースの id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: >-
        Authorization ヘッダーが無い、または Bearer 方式でない（authentication_required）か、
        Bearer 方式で資格情報が無い・壊れている・期限切れ・利用者が退会済み（invalid_token。どれに当たったかは区別しない。機能一覧 1.4）。
        アクセストークンはリフレッシュ（/auth/refresh）で取り直す。仕様の形の検証（400）はトークンの確認より先に行う（仕様で書けない検証は後に行う）
      headers:
        WWW-Authenticate:
          description: >-
            Authorization ヘッダーが無い・Bearer 方式でないときは `Bearer`、Bearer 方式で資格情報が無い・使えないときは `Bearer error="invalid_token"`（RFC 6750 3.1）
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: >-
        入力が仕様に合わない（validation_failed）か、本体を JSON として読めない（invalid_body）。
        どちらも送られた値を応答に載せない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowed:
      description: 仕様に無いメソッド（method_not_allowed）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: 本体が大きすぎる（payload_too_large。上限は Nest の JSON の読み取りの既定の 100kb）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: 仕様に無いメディア型の本体（unsupported_media_type）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: レート制限を超えた（too_many_requests。上限はパスの説明に書く）
      headers:
        Retry-After:
          description: 再び受け付けるまでの秒数
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: >-
        見つからない（not_found）。所属していないワークスペースは、存在の有無を区別せずこれになる（機能一覧 2.1・1.4）。
        本体は状態コードの本体であり、権限の有無を述べない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 想定外の失敗（internal_error）。例外のメッセージは載せない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          description: >-
            エラーの種類。api が返す値はこの列挙だけであり、api と web は生成した型で同じ列挙を使う
            （綴りを誤ると型検査で落ちる）
          type: string
          enum:
            - validation_failed
            - invalid_body
            - not_found
            - method_not_allowed
            - payload_too_large
            - unsupported_media_type
            - too_many_requests
            - user_id_taken
            - registration_disabled
            - invalid_credentials
            - authentication_required
            - invalid_token
            - csrf_rejected
            - owner_only
            - invitee_not_found
            - already_invited
            - already_member
            - owner_cannot_leave
            - password_mismatch
            - owner_cannot_delete_account
            - pin_limit_reached
            - channel_name_taken
            - not_a_channel_member
            - already_channel_member
            - channel_archived
            - channel_not_private
            - channel_not_archived
            - not_message_author
            - dm_with_self
            - dm_counterpart_not_found
            - dm_counterpart_unavailable
            - reaction_limit_reached
            - unsupported_file_type
            - file_too_large
            - upload_not_received
            - upload_in_progress
            - attachment_unavailable
            - request_rejected
            - internal_error
        message:
          type: string
        errors:
          description: 入力の検証で落ちた箇所。送られた値は含めない
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - path
              - message
            properties:
              path:
                type: string
              message:
                type: string
...
    UploadRequest:
      description: >-
        アップロードの発行の要求（機能一覧 11.1）。どれも申告であり、形式と大きさは確定で中身から検証する。
        キー・パス・URL は指定できない（サーバーが組み立てる）
      type: object
      additionalProperties: false
      required:
        - fileName
        - contentType
        - size
      properties:
        fileName:
          description: 元のファイル名（1〜255 文字）。保存名はこれを置き換えて作る
          type: string
          minLength: 1
          maxLength: 255
        contentType:
          description: >-
            申告する Content-Type。許可リスト（packages/shared の UPLOAD_FORMATS）の値でなければ 422 unsupported_file_type。
            署名付き URL の PUT には、この値をそのまま Content-Type に付ける
          type: string
          minLength: 1
          maxLength: 255
        size:
          description: 申告する大きさ（バイト）。種別の上限を超えれば 422 file_too_large
          type: integer
          minimum: 1
    UploadTicket:
      description: 発行したアップロード（機能一覧 11.1）
      type: object
      additionalProperties: false
      required:
        - uploadId
        - uploadUrl
        - uploadHeaders
        - expiresAt
      properties:
        uploadId:
          description: アップロードの識別子。確定の要求に使う
          type: string
          format: uuid
        uploadUrl:
          description: 本体を PUT する署名付き URL（宛先は隔離用のキー）
          type: string
        uploadHeaders:
          description: PUT に付けるヘッダー。どちらも署名に含まれ、変えたり外したりすると断られる
          type: object
          additionalProperties: false
          required:
            - Content-Type
            - If-None-Match
          properties:
            Content-Type:
              type: string
            If-None-Match:
              type: string
              enum:
                - '*'
        expiresAt:
          description: 署名付き URL の有効期限（発行から 5 分）。この時刻までに PUT を始める
          type: string
          format: date-time
...
    Attachment:
      description: 添付ファイル（機能一覧 11.1・11.2）
      type: object
      additionalProperties: false
      required:
        - id
        - fileName
        - contentType
        - kind
        - size
        - url
      properties:
        id:
          description: アップロードの識別子
          type: string
          format: uuid
        fileName:
          description: 利用者が付けた元のファイル名（表示用。保存名とは違う）
          type: string
        contentType:
          description: 配信の Content-Type（検証した形式からサーバーが決めた値。テキスト系は text/plain; charset=utf-8）
          type: string
        kind:
          description: 種別（画像は表示し、動画は #t=30 の位置を表示し、それ以外はリンクにする。機能一覧 11.1）
          type: string
          enum:
            - image
            - video
            - document
            - archive
        size:
          description: 大きさ（バイト）
          type: integer
          minimum: 1
        url:
          description: >-
            配信 URL のパス（/files/workspace/{workspaceId}/channel/{channelId}/{id}/{保存名}）。CloudFront の署名付き Cookie で取得する（機能一覧 11.2）
          type: string
...